Further tidy up. Introduce concept of a "subscribable".
