+++ src/DomainBlocks.Core/Projections/Experimental/Builders/EventStreamConsumersBuilderExtensions.cs
@@ -2,7 +2,7 @@ using DomainBlocks.Core.Subscriptions.Builders;
 
 namespace DomainBlocks.Core.Projections.Experimental.Builders;
 
-public static class EventStreamSubscribersBuilderExtensions
+public static class EventStreamConsumersBuilderExtensions
 {
     public static ProjectionsBuilder<TEvent, TPosition> ProjectTo<TEvent, TPosition>(
         this EventStreamConsumersBuilder<TEvent, TPosition> consumersBuilder)
