+++ src/DomainBlocks.Core/Subscriptions/IEventStreamSubscriber.cs
@@ -1,12 +1,13 @@
 namespace DomainBlocks.Core.Subscriptions;
 
-public interface IEventStreamSubscriber<out TEvent, TPosition> where TPosition : struct
+public interface IEventStreamSubscriber<in TEvent, in TPosition>
 {
-    Task<IDisposable> Subscribe(
-        TPosition? fromPositionExclusive,
-        Func<CancellationToken, Task> onCatchingUp,
-        Func<TEvent, TPosition, CancellationToken, Task> onEvent,
-        Func<CancellationToken, Task> onLive,
-        Func<SubscriptionDroppedReason, Exception?, CancellationToken, Task> onSubscriptionDropped,
-        CancellationToken cancellationToken);
+    Task OnCatchingUp(CancellationToken cancellationToken);
+
+    Task OnEvent(TEvent @event, TPosition position, CancellationToken cancellationToken);
+
+    Task OnLive(CancellationToken cancellationToken);
+
+    Task OnSubscriptionDropped(
+        SubscriptionDroppedReason reason, Exception? exception, CancellationToken cancellationToken);
 }
