+++ src/DomainBlocks.Core/Subscriptions/QueueNotificationType.cs
@@ -1,0 +1,10 @@
+namespace DomainBlocks.Core.Subscriptions;
+
+internal enum QueueNotificationType
+{
+    CatchingUp,
+    Event,
+    Live,
+    SubscriptionDropped,
+    CheckpointTimerElapsed
+}
